backend/requirements.txt
langchain>=0.2.0
langchain-google-genai>=0.0.8
langchain-mongodb>=0.1.5
google-generativeai>=0.5.2
pymongo>=4.6.0
python-dotenv>=1.0.1
numpy>=1.24.0
Flask>=3.0.0
Flask-Cors>=4.0.0
